Fix Dockerfile and added a client
I fixed the Dockerfile to now work off our custom base.
Also, I created a client for the tts service so it's easier to use. The tts_command was changed to start_server and we now have the command say to send text.
Lastly, I added an install_dependencies file to install the software dependencies, this way we can install on a non-docker machine if necessary.